Add 2d camera

+++ crates/tubereng_renderer/src/texture.rs
@@ -71,8 +71,8 @@ pub struct Descriptor<'a> {
 
 #[derive(Debug, Clone)]
 pub struct Rect {
-    pub x: u32,
-    pub y: u32,
-    pub width: u32,
-    pub height: u32,
+    pub x: f32,
+    pub y: f32,
+    pub width: f32,
+    pub height: f32,
 }
